ACE-126 Fix session issue with SSR (#150)
