+++ src/app/rounds/[roundId]/round-component.tsx
@@ -6,9 +6,7 @@ import IssueDisplay from '@aces/components/issues/issue-display'
 import LoadingRound from '@aces/components/rounds/loading-round'
 import { RoundSidebar } from '@aces/components/rounds/sidebar'
 import { View } from '@aces/interfaces/view'
-import { useCsrfToken } from '@aces/lib/hooks/auth/use-csrf-token'
 import useCurrentUser from '@aces/lib/hooks/auth/use-current-user'
-import useMigrateCookie from '@aces/lib/hooks/auth/use-migrate-cookie'
 import useRegisterViewer from '@aces/lib/hooks/auth/use-register-viewer'
 import WebSocketConnection from '@aces/lib/socket/web-socket-connection'
 import { useWebSocket } from '@aces/lib/socket/web-socket-provider'
@@ -20,8 +18,6 @@ interface RoundComponentProps {
 }
 
 function RoundComponent({ roundId, views }: RoundComponentProps): React.ReactElement {
-  const { csrfToken } = useCsrfToken()
-  useMigrateCookie(csrfToken)
   const { user, isLoading: isUserLoading } = useCurrentUser()
   useRegisterViewer({ roundId }, isUserLoading ? undefined : user)
   const { isConnected } = useWebSocket()
